+++ MaxMind.MinFraud/Exception/AuthenticationException.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Runtime.Serialization;
 
 namespace MaxMind.MinFraud.Exception
 {
@@ -30,18 +29,6 @@ namespace MaxMind.MinFraud.Exception
         /// <param name="message">Exception message.</param>
         /// <param name="innerException">The underlying exception that caused this one.</param>
         public AuthenticationException(string message, System.Exception innerException) : base(message, innerException)
-        {
-        }
-
-        /// <summary>
-        ///     Constructor for deserialization.
-        /// </summary>
-        /// <param name="info"></param>
-        /// <param name="context"></param>
-#if NET8_0_OR_GREATER
-        [Obsolete(DiagnosticId = "SYSLIB0051")]
-#endif
-        protected AuthenticationException(SerializationInfo info, StreamingContext context) : base(info, context)
         {
         }
     }