+++ MaxMind.MinFraud/Exception/HttpException.cs
@@ -3,7 +3,6 @@
 using System;
 using System.IO;
 using System.Net;
-using System.Runtime.Serialization;
 
 #endregion
 
@@ -73,37 +72,6 @@ namespace MaxMind.MinFraud.Exception
         {
             HttpStatus = httpStatus;
             Uri = uri;
-        }
-        /// <summary>
-        ///     Constructor for deserialization.
-        /// </summary>
-        /// <param name="info"></param>
-        /// <param name="context"></param>
-#if NET8_0_OR_GREATER
-        [Obsolete(DiagnosticId = "SYSLIB0051")]
-#endif
-        protected HttpException(SerializationInfo info, StreamingContext context) : base(info, context)
-        {
-            HttpStatus = (HttpStatusCode)(info.GetValue("MaxMind.MinFraud.Exception.HttpException.HttpStatus", typeof(HttpStatusCode))
-                ?? throw new SerializationException("Unexpected null HttpStatus value"));
-            Uri = (Uri)(info.GetValue("MaxMind.MinFraud.Exception.HttpException.Uri", typeof(Uri))
-                ?? throw new SerializationException("Unexpected null Uri value"));
-        }
-
-        /// <summary>
-        ///     Method to serialize data.
-        /// </summary>
-        /// <param name="info"></param>
-        /// <param name="context"></param>
-#if NET8_0_OR_GREATER
-        [Obsolete(DiagnosticId = "SYSLIB0051")]
-#endif
-        public override void GetObjectData(SerializationInfo info, StreamingContext context)
-        {
-            base.GetObjectData(info, context);
-
-            info.AddValue("MaxMind.MinFraud.Exception.HttpException.HttpStatus", HttpStatus, typeof(HttpStatusCode));
-            info.AddValue("MaxMind.MinFraud.Exception.HttpException.Uri", Uri, typeof(Uri));
         }
 
         /// <summary>